requirements.txt
# Core LangChain packages
langchain-core==0.3.69
langchain==0.3.26
langchain-community==0.3.27
langchain-groq==0.2.1

# For your Ollama embeddings in vector.py
langchain-ollama==0.2.2

# Web frameworks
streamlit==1.40.2
fastapi==0.115.6
uvicorn==0.32.1

# Environment
python-dotenv==1.0.1

# Document processing
beautifulsoup4==4.12.3
pypdf==5.1.0

# Vector stores
faiss-cpu==1.9.0
chromadb==0.5.23

# External APIs
groq==0.13.0

# Research tools (for ArxivQueryRun)
arxiv==2.1.3
wikipedia==1.4.0

# Additional dependencies
requests==2.32.3
xmltodict==0.13.0
sentence-transformers==3.3.1

# Optional
langchainhub==0.1.21
cassio==0.1.9